Oh christ. So: if you caused your own room to be resolved, it would add the puppet and it would be off looking shit up in the database while you entered. And that would end up trampling memories.
So it's a bit hackier, but now we check to see if the puppet should be loaded
when someone enters, and we don't re-add the puppet if the owner leaves and
the room would be empty. (Even though it's practically "free" at that
point, because we've got the avatar and memories loaded.)

But this also has the benefit of not showing the puppet when someone
is doing a "live view" of a room. Aw fuck, but I just realized it will
show the puppet in the live view if a non-owner was present and left.. I think.