Bugfix: fix the menu that's popped when a user's name is in a text link, if the user has "/" in their name.
We use "/" as a separator when parsing commands. It's a good separator.
Nice and readable. It would be a shame to lose it. So continue to use it
by default, but if the first character of the command is \uFFFC, then
chop it off and use it as the separator.

This character is already used to delimit linky text, but that regex is
greedy and terminates with \uFFFD, so we can mix these in. It's a bit of
a mess and needs all this explanation, but I don't have any better idea.