As far as I can tell, we have not been persisting entry vectors for embedded plays for months, probably even longer. Every day, thousands of permaguests are created with no vector information associated.
Our code goes to great lengths to catch all the places in which a new visitor has arrived and create the association then and there. When arriving through GWT, the visitor is immediately associated with a vector as soon as they hit the landing page.

For embeds, however, there is no GWT, and the Flash client creates its own visitorId which is then passed in through the Credentials object. There is code in GameSession and MsoySession to create the association there -- unfortunately, it relies on memobj.visitorId to be null for just-created guests, and that is never true. All permaguests are created with a visitorId. I don't know how long ago it is that this test worked.

I have been pondering the best way to fix the problem. Probably, we should have the Flash client tell us if the visitorId it's including in the login credentials is freshly created. For now, though, we do a database lookup and create the association if one doesn't already exist. I want to see how well this patch does before I invest more time into this gigantic rabbit hole.