MoneyTransactionRecord is the single biggest overall database resource hog for Whirled, and individual queries on it sometimes run for minutes. This refactor splits it into three tables, one for each currency.
The main problem with the combined table, apart from its needlessly large size, is the heterogenous distribution of its rows. The COIN rows are very dense, but all recent (they are purged quite quickly). The BLING and BAR rows, by contrast, are archived indefinitely. The query optimizer uses sampling techniques to figure out approximately how data is distributed over the row, and it's very confused by the fact that two such different data sets share the same table. The result is suboptimal queries that almost certainly uses the wrong indexes from time to time.

The only drawback to separate tables is for queries that return e.g. chronological views across all currencies. The only spot where we still do this is for the item transaction view, which is a relatively rare thing to be looking at. I wrote a compatibility method that pulls from all three tables and returns the result.

Because of the size of this table, I am going to do the data migration by hand, so there is no automated migration of data from MoneyTransactionRecord to the three new tables. I can't see this being much of an issue since I'm the only remaining developer. :)