OFFS. We use GWTCanvas to display game stat graphs. I so should have just used Google's graphing services. Indeed, I'd rewrite things now, but getting the amazing four hundred billion moving parts back into place to have an operational Whirled development environment would dwarf that effort by about tenfold.
I looked for a more modern Canvas widget for GWT, but everything seems at least
one or two years old. Fortunately the GWTCanvas from gwt-incubator doesn't seem
to refer to any of the now-incompatible-with-GWT-2.1.0 code, so we can just
keep using it.

Note to self: don't introduce dependencies on wacky fucking libraries, they're
not going to be maintained, and they'll be a pain in the ass as you try to keep
up with the evolution of your core platform.