Consolidate the way we style links.
Half the links on the site were styled as only underlining when you hover over them, the other half
were black and always underlined. Those links are now Whirled-blue and underline only on hovering.

I'll be watching dev for a few days to make sure this looks right everywhere, but so far it appears
to be a big improvement!