Disable profile interest linking.
This is currently broken, and looks like it's been that way for a while. Clicking on someone's
interest of eg. "spelunking" takes you to a search for players named spelunking, which is useless. I
swear I remember this used to instead search for other people also interested in spelunking. I'll
look into it later.