Back-populate missing bling pool distributions.
For whatever reason, we had a bunch of old transaction records with IDs greater than the sequence
that generates them. This was causing the sequence to occasionally generate duplicate keys and fail
to record transactions. For item sales, this has resulted in missing transactions, but luckily we
still actually credited those accounts... For games, we've been missing payments.

This chunk of code should recalculate and distribute bling to all games that have missing
transactions, from as far back as we have game play logs.