Re-use the selected tab text color for the links on the title bar.
This leaves the title bar header text color, which is still hard coded to dark blue at the moment.
Instead of adding tons more theme fields, I'm starting to think we should go Myspace-style and allow
players to upload CSS files that override ours. We would still have the same theme fields as now,
but CSS can be used to skin the bits we never supported. This is mainly to expedite the construction
of Whirled's DJ destination site, but it would be very interesting to see what players do with it as
well.