Remove the close button on the Flash, in all client modes.
We could make this a home button instead, as Daniel suggested. I think we have enough ways to get
back home though, both in the Flash control bar and GWT.

On navigation: Looking at the "Go..." menu in Flash, the most important thing there is the recent scenes
list IMO, and it's buried in a submenu. We should remove the other stuff and make that the top level
menu...