Fix for shadow fringe rendering glitchiness. I blame the Flash player.
This was working when I wrote it a couple weeks ago, but lately started looking like this:
https://share.catdap.com/UmNFXJ4oq4aSspciy1QMpA.png

Note the masking problems with the door within the scene, which is a completely separate issue but
included for laughs. The shadow fringe being an ugly line was happening on Windows too, luckily this
patch works around it. Maybe.

The door masking problem only seems to happen when the door is occluded a certain way. I'm going to
ignore it.