Disable Cloudfront URL signing.
This still needs a lot of work before it's ready to be enabled. In the meantime, the frequently
updating URL params are breaking caching and probably costing us more bandwidth than hotlinkers. At
the least, this will improve site speed for our players as they don't have to be redownloading media
all the time.

If anyone ever wants to finish off Cloudfront URL signing, there's a discussion between Zell and I
that I should forward you. Briefly, there needs to be a way to have Flash use a signed
crossdomain.xml, and the Expires param has to be way more granular so it doesn't break caching.