Disable a bunch of emailing.
I'd prefer not to, but our ability to send email at all is going to be
suspended because the bounce rate is too high. Most likely the main culprit is
mail going out when a Whirled Mail is sent, but the other mails are problematic
as well.

This way we can at least continue to send critical emails like password reset
and registration confirmation rather than having AWS cut us off and being
totally screwed.