Remove threerings library dependency.
Replace it with the half-dozen smaller libraries it was split into, the source
for most of which is available on the internets. ooo-messaging is not
available, but I'll look into releasing that, or at least getting a binary into
the same place that all the other legacy Whirled depends are.